requirements.txt
torch==2.2.1
torchvision==0.17.1
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
mlflow==2.12.1
dvc==3.49.0
pytest==8.1.1
httpx==0.27.0
scikit-learn==1.4.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
prometheus-client==0.20.0
pydantic==2.7.0
